sw_core/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==3.2.20
djangorestframework==3.13.1
dynamic-rest==2.1.2
frozenlist==1.4.0
greenlet==2.0.2
hashids==1.3.1
idna==3.4
inflection==0.5.1
kombu==5.3.1
multidict==6.0.4
mysql-connector==2.2.9
mysql-connector-python==8.1.0
mysqlclient==2.2.0
playwright==1.37.0
prompt-toolkit==3.0.39
protobuf==4.21.12
pyee==9.0.4
python-dateutil==2.8.2
pytz==2023.3
redis==5.0.0
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
